changed the test file name changed the Epexcted output as well
